chore: de-duplicate integration tests
